Added glm as a history-less submodule.
The actual glm repo is >80MB and I didn't want the history anyways. This
is the best compromise. Now it's just referencing a repo with newest glm
but in one commit. If I want to update glm I just need to copy all the
new files in. Not great but acceptable.